Zavrsne izmjene - dizajn, zadnje promjene na formama

+++ Software/RestoranAPK/FormNarudzbe.cs
@@ -17,6 +17,7 @@ namespace Funkcionalnost_prijave
         public List<Order> ListaNarudzbi { get; set; }
         public User LogiranKorisnik { get; set; }
         public Order OdabranaNarudzba { get; set; }
+        public string red;
         public FormNarudzbe(User korisnik)
         {
             InitializeComponent();
@@ -87,26 +88,41 @@ namespace Funkcionalnost_prijave
 
         private void buttonObrisiNarudzbu_Click(object sender, EventArgs e)
         {
-            Order odabrana = dataGridViewNarudzbe.CurrentRow.DataBoundItem as Order;
-            using (var context = new EntitiesOrder())
+            if (dataGridViewNarudzbe.CurrentRow == null)
             {
-                foreach (var item in context.Carts)
-                {
-                    if (item.Narudzba == odabrana.ID)
-                    {
-                        context.Carts.Remove(item);
-                    }
-                }
-                foreach (var item in context.Orders)
-                {
-                    if (item.ID == odabrana.ID)
-                    {
-                        context.Orders.Remove(item);
-                    }
-                }
-                context.SaveChanges();
+                red = "";
             }
-            Osvjezi();
+            else
+            {
+                red = dataGridViewNarudzbe.CurrentRow.ToString();
+            }
+            if (BibliotekeVanjske.ValidacijaUnosa.ProvjeriOdabirReda(red) == "")
+            {
+                Order odabrana = dataGridViewNarudzbe.CurrentRow.DataBoundItem as Order;
+                using (var context = new EntitiesOrder())
+                {
+                    foreach (var item in context.Carts)
+                    {
+                        if (item.Narudzba == odabrana.ID)
+                        {
+                            context.Carts.Remove(item);
+                        }
+                    }
+                    foreach (var item in context.Orders)
+                    {
+                        if (item.ID == odabrana.ID)
+                        {
+                            context.Orders.Remove(item);
+                        }
+                    }
+                    context.SaveChanges();
+                }
+                Osvjezi();
+            }
+            else
+            {
+                MessageBox.Show(BibliotekeVanjske.ValidacijaUnosa.ProvjeriOdabirReda(red));
+            }
         }
 
         private void buttonPovratak_Click(object sender, EventArgs e)
@@ -121,19 +137,35 @@ namespace Funkcionalnost_prijave
 
         private void buttonPromijeniStatus_Click(object sender, EventArgs e)
         {
-            Order odabrana = dataGridViewNarudzbe.CurrentRow.DataBoundItem as Order;
-            using (var context = new EntitiesOrder())
+            if (dataGridViewNarudzbe.CurrentRow == null)
             {
-                foreach (var item in context.Orders)
-                {
-                    if(item.ID == odabrana.ID)
-                    {
-                        item.Status = comboBoxStatus.Text;
-                    }
-                }
-                context.SaveChanges();
+                red = "";
             }
-            Osvjezi();
+            else
+            {
+                red = dataGridViewNarudzbe.CurrentRow.ToString();
+            }
+            if (BibliotekeVanjske.ValidacijaUnosa.ProvjeriOdabirReda(red) == "")
+            {
+                Order odabrana = dataGridViewNarudzbe.CurrentRow.DataBoundItem as Order;
+                using (var context = new EntitiesOrder())
+                {
+                    foreach (var item in context.Orders)
+                    {
+                        if (item.ID == odabrana.ID)
+                        {
+                            item.Status = comboBoxStatus.Text;
+                        }
+                    }
+                    context.SaveChanges();
+                }
+                Osvjezi();
+            }
+            else
+            {
+                MessageBox.Show(BibliotekeVanjske.ValidacijaUnosa.ProvjeriOdabirReda(red));
+            }
+          
         }
 
         private void buttonDodajNarudzbu_Click(object sender, EventArgs e)
@@ -145,18 +177,54 @@ namespace Funkcionalnost_prijave
 
         private void buttonPrikaziNarudzbu_Click(object sender, EventArgs e)
         {
-            OdabranaNarudzba = dataGridViewNarudzbe.CurrentRow.DataBoundItem as Order;
-            Hide();
-            using (var forma = new FormPrikaziNarudzbu(LogiranKorisnik,OdabranaNarudzba))
+            if (dataGridViewNarudzbe.CurrentRow == null)
             {
-                forma.ShowDialog();
+                red = "";
             }
-            Close();
+            else
+            {
+                red = dataGridViewNarudzbe.CurrentRow.ToString();
+            }
+            if (BibliotekeVanjske.ValidacijaUnosa.ProvjeriOdabirReda(red) == "")
+            {
+                OdabranaNarudzba = dataGridViewNarudzbe.CurrentRow.DataBoundItem as Order;
+                Hide();
+                using (var forma = new FormPrikaziNarudzbu(LogiranKorisnik, OdabranaNarudzba))
+                {
+                    forma.ShowDialog();
+                }
+                Close();
+            }
+            else
+            {
+                MessageBox.Show(BibliotekeVanjske.ValidacijaUnosa.ProvjeriOdabirReda(red));
+            }
         }
 
         private void FormNarudzbe_HelpRequested(object sender, HelpEventArgs hlpevent)
         {
             Pomoc();
+        }
+
+        private void dataGridViewNarudzbe_CellContentClick(object sender, DataGridViewCellEventArgs e)
+        {
+
+        }
+
+        private void label4_Click(object sender, EventArgs e)
+        {
+            Hide();
+            using (var forma = new FormPrijavljenZaposlenik(LogiranKorisnik))
+            {
+                forma.ShowDialog();
+            }
+            Close();
+
+        }
+
+        private void label2_Click(object sender, EventArgs e)
+        {
+            Close();
         }
     }
 }
